Rename seed view and controller
