+++ server/test/controllers/dev/DevToolsControllerTest.java
@@ -1,4 +1,4 @@
-package controllers.dev.seeding;
+package controllers.dev;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static play.mvc.Http.Status.NOT_FOUND;
@@ -21,10 +21,10 @@ import play.mvc.Result;
 import play.test.Helpers;
 import repository.VersionRepository;
 
-public class DevDatabaseSeedControllerTest {
+public class DevToolsControllerTest {
 
   private Optional<Application> maybeApp = Optional.empty();
-  private DevDatabaseSeedController controller;
+  private DevToolsController controller;
   private VersionRepository versionRepo;
   private SyncCacheApi programsByVersionCache;
 
@@ -47,7 +47,7 @@ public class DevDatabaseSeedControllerTest {
 
     // Seed the fake data.
     result = controller.seedPrograms();
-    assertThat(result.redirectLocation()).hasValue(routes.DevDatabaseSeedController.index().url());
+    assertThat(result.redirectLocation()).hasValue(routes.DevToolsController.index().url());
     assertThat(result.flash().get(FlashKey.SUCCESS)).hasValue("The database has been seeded");
     result = controller.index(fakeRequest());
     assertThat(result.status()).isEqualTo(OK);
@@ -60,7 +60,7 @@ public class DevDatabaseSeedControllerTest {
 
     // Clear the data.
     result = controller.clear();
-    assertThat(result.redirectLocation()).hasValue(routes.DevDatabaseSeedController.index().url());
+    assertThat(result.redirectLocation()).hasValue(routes.DevToolsController.index().url());
     assertThat(result.flash().get(FlashKey.SUCCESS)).hasValue("The database has been cleared");
     result = controller.index(fakeRequest());
     assertThat(result.status()).isEqualTo(OK);
@@ -85,7 +85,7 @@ public class DevDatabaseSeedControllerTest {
 
     // Seed the fake data.
     result = controller.seedPrograms();
-    assertThat(result.redirectLocation()).hasValue(routes.DevDatabaseSeedController.index().url());
+    assertThat(result.redirectLocation()).hasValue(routes.DevToolsController.index().url());
     assertThat(result.flash().get(FlashKey.SUCCESS)).hasValue("The database has been seeded");
     result = controller.index(fakeRequest());
     assertThat(result.status()).isEqualTo(OK);
@@ -142,7 +142,7 @@ public class DevDatabaseSeedControllerTest {
                 .in(mode)
                 .configure("version_cache_enabled", true)
                 .build());
-    controller = maybeApp.get().injector().instanceOf(DevDatabaseSeedController.class);
+    controller = maybeApp.get().injector().instanceOf(DevToolsController.class);
     BindingKey<SyncCacheApi> versionProgramsKey =
         new BindingKey<>(SyncCacheApi.class).qualifiedWith(new NamedCacheImpl("version-programs"));
     programsByVersionCache = maybeApp.get().injector().instanceOf(versionProgramsKey.asScala());
